add share for crash detection tool
