+++ Sentinel/Classes/DatabaseImportExport/DatabaseImportExportView.swift
@@ -52,71 +52,3 @@ struct DatabaseImportExportView: View {
         }
     }
 }
-
-// MARK: - MacOS share
-
-#if os(macOS)
-struct SharingsPicker: NSViewRepresentable {
-    @Binding var isPresented: Bool
-    let sharingItems: [Any]
-    let didFinish: () -> Void
-
-    func makeNSView(context: Context) -> NSView {
-        NSView()
-    }
-
-    func updateNSView(_ nsView: NSView, context: Context) {
-        if isPresented {
-            let picker = NSSharingServicePicker(items: sharingItems)
-            picker.delegate = context.coordinator
-
-            DispatchQueue.main.async {
-                picker.show(relativeTo: .zero, of: nsView, preferredEdge: .minY)
-            }
-        }
-    }
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(picker: self)
-    }
-
-    class Coordinator: NSObject, NSSharingServicePickerDelegate {
-        let picker: SharingsPicker
-
-        init(picker: SharingsPicker) {
-            self.picker = picker
-        }
-
-        func sharingServicePicker(_ sharingServicePicker: NSSharingServicePicker, didChoose service: NSSharingService?) {
-            sharingServicePicker.delegate = nil
-            picker.isPresented = false
-            picker.didFinish()
-        }
-    }
-}
-#endif
-
-// MARK: - iOS share
-
-#if os(iOS)
-struct ActivityViewController: UIViewControllerRepresentable {
-
-    @Environment(\.presentationMode) var presentationMode
-
-    let activityItems: [Any]
-    let applicationActivities: [UIActivity]?
-    let didFinish: () -> Void
-
-    func makeUIViewController(context: UIViewControllerRepresentableContext<ActivityViewController>) -> UIActivityViewController {
-        let controller = UIActivityViewController(activityItems: activityItems, applicationActivities: applicationActivities)
-        controller.completionWithItemsHandler = { (activityType, completed, returnedItems, error) in
-            self.presentationMode.wrappedValue.dismiss()
-            didFinish()
-        }
-        return controller
-    }
-
-    func updateUIViewController(_ uiViewController: UIActivityViewController, context: UIViewControllerRepresentableContext<ActivityViewController>) {}
-
-}
-#endif
